fix: table at cursor position showing markdown
Desktop already had logic to to move the cursor out of a table when switching notes to prevent the table showing as raw markdown when switching to a note where the saved cursor position was in a table, but the logic didn't work on mobile because on mobile the content script doesn't load until you switch to edit mode.

There was also a remaining issue on desktop where the first note loaded after cold start could still have the issue because the note "switch" logic wouldn't fire in this case.

Fix both issues by triggering the move cursor out of table logic on content script load.

This works for the note switch scenario on mobile because the content script is always loaded fresh when switching to a note and then switching to edit mode.

On desktop the content script is only loaded when first accessing a note so it addresses the cold start scenario and doesn't change behavior otherwsie.